SNC-6236. Changes compatible with Qt5

Closes SNC-6236

See merge request desktop/desktop!4116

+++ src/MEGASync/control/Utilities.h
@@ -2,6 +2,7 @@
 #define UTILITIES_H
 
 #include "megaapi.h"
+#include "QtMetaEnumUtils.h"
 #include "ThreadPool.h"
 
 #include <QDesktopServices>
@@ -626,7 +627,7 @@ public:
     static void sleepMilliseconds(unsigned int milliseconds);
 
     // Compute the part per <ref> of <part> from <total>. Defaults to %
-    static int partPer(long long part, long long total, uint ref = 100);
+    static int partPer(long double part, long double total, uint ref = 100);
 
     static QString getFileHash(const QString& filePath);
